Move evm ts to evm directory
