+++ evm/Dockerfile
@@ -4,9 +4,11 @@ WORKDIR /app
 COPY foundry.toml foundry.toml
 COPY lib lib
 COPY src src
+COPY ts ts
 
 RUN FOUNDRY_PROFILE=prod forge build
 
 FROM scratch AS foundry-export
 
 COPY --from=builder /app/out .
+COPY --from=builder /app/ts ts
