+++ evm/ts/.gitignore
@@ -1,0 +1,3 @@
+dist
+ethers-ci-contracts
+contracts.json
