+++ evm/ts/scripts/readVersion.ts
