+++ evm/ts/src/ethers-contracts/1_0_0/factories/NttManager__factory.ts
