+++ evm/ts/src/ethers-contracts/1_1_0/factories/NttManager__factory.ts
