+++ evm/ts/src/ntt.ts
