+++ sdk/definitions/tsconfig.cjs.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../tsconfig.cjs.json",
+  "extends": "../../tsconfig.cjs.json",
   "include": ["src"],
   "compilerOptions": {
     "outDir": "dist/cjs",
