+++ sdk/definitions/tsconfig.esm.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../tsconfig.esm.json",
+  "extends": "../../tsconfig.esm.json",
   "include": ["src"],
   "compilerOptions": {
     "outDir": "dist/esm",
