+++ sdk/examples/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../tsconfig.esm.json",
+  "extends": "../../tsconfig.esm.json",
   "include": ["src/**/*.ts", "src/**/*.json"],
   "compilerOptions": {
     "rootDir": "src",
